Added JS Coin Animation and DIV removal
refactored the Coin DIV creation to work for both Home and Visitor in 1 fuction. Created the DIV removal to occur at the end of the animation step. adjusted the CSS for the 2 parameter function call.  Added gitignore file and added the 'codeTestFunctions.js to be ignore - I can dump test and trial code in there in case I want to refer to later